+++ wazo_auth/database/queries/session.py
@@ -14,13 +14,16 @@ class SessionDAO(PaginatorMixin, BaseDAO):
 
     column_map = {'mobile': Session.mobile}
 
-    def list_(self, tenant_uuids=None, **kwargs):
+    def list_(self, tenant_uuids=None, user_uuid=None, **kwargs):
         filter_ = text('true')
         if tenant_uuids is not None:
             if not tenant_uuids:
                 return []
 
-            filter_ = Session.tenant_uuid.in_(tenant_uuids)
+            filter_ = and_(filter_, Session.tenant_uuid.in_(tenant_uuids))
+
+        if user_uuid is not None:
+            filter_ = and_(filter_, Token.auth_id == str(user_uuid))
 
         with self.new_session() as s:
             query = s.query(Session, Token).join(Token).filter(filter_)